feat: Introduce DTOs and mappers for user and product services, standardize global exception handling, and update service configurations.

+++ order-service/src/main/java/com/example/orderservice/dto/OrderResponse.java
@@ -1,15 +1,16 @@
 package com.example.orderservice.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record OrderResponse(
-    Long id,
-    Long userId,
-    String userName,
-    Long productId,
-    String productName,
-    Integer quantity,
-    Double totalAmount,
-    String status,
-    LocalDateTime orderDate
-) {}
+        Long id,
+        Long userId,
+        String userName,
+        Long productId,
+        String productName,
+        Integer quantity,
+        BigDecimal totalAmount,
+        String status,
+        LocalDateTime orderDate) {
+}
