+++ order-service/src/main/java/com/example/orderservice/entity/Order.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -17,14 +18,14 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    
+
     private Long userId;
     private Long productId;
     private Integer quantity;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private String status;
     private LocalDateTime orderDate;
-    
+
     @PrePersist
     public void prePersist() {
         orderDate = LocalDateTime.now();
